requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
requests==2.28.2
beautifulsoup4==4.11.1
python-dotenv==1.0.0
